audio support support for shockspot altered playback speeds

+++ pkg/ffmpeg/types.go
@@ -64,7 +64,7 @@ type FFProbeStream struct {
 		VisualImpaired  int `json:"visual_impaired"`
 	} `json:"disposition"`
 	Duration          string `json:"duration"`
-	DurationTs        int    `json:"duration_ts"`
+	DurationTs        int64    `json:"duration_ts"`
 	HasBFrames        int    `json:"has_b_frames,omitempty"`
 	Height            int    `json:"height,omitempty"`
 	Index             int    `json:"index"`
